+++ components/SocialMediaIcons.tsx
@@ -14,11 +14,6 @@ export const SocialMediaIcons = () => {
         bgColor="#fff"
         style={{ width: 30, height: 30 }}
       />
-      <SocialIcon
-        url="https://www.telegram.me/juliancwirko"
-        bgColor="#fff"
-        style={{ width: 30, height: 30 }}
-      />
     </Box>
   );
 };