+++ config/dappCustoms.ts
@@ -1,0 +1,4 @@
+export const avatarIdUrl = (address: string) =>
+  `https://id.maiar.com/users/photos/profile/${address}`;
+
+export const customIPFSGateway = 'https://nftstorage.link/ipfs/';